Solve ChordSheetEditor issues
